install: move three optional installs into install_requires
with this the tests install is simplified because with `pip install '.[tests]'`
all necessary packages are installed. It is not enough to only move those
packages into 'tests' and keep the optional keys for backward compatibility
reasons, because if the package is installed into production with the optional
keys it want work because they only would be installed for tests reason.
therefore those package had to be moved to install_requires.