perf(models): change `extra_data` to `JSONB`
* `JSONB` is a newer and more efficient format for storing JSON in
PostgreSQL. Compared to `JSON`, it offers significantly faster querying
and indexing support (but is very slightly slower to input). The JSON is
decomposed into a binary format rather than being stored in plaintext.

* From a high-level perspective, the access of the data is unchanged in
Python so this is not a breaking change. However, the new column type
supports more querying operations which could be used in the future.

* Migrating the column type is quite easy for small tables but can cause
performance and stability issues for larger ones. Therefore, a default
Alembic migration has been included for the vast majority of use cases,
as well as a step-by-step alternative guide for larger instances.
Instances with more than ~50k rows in `oauthclient_remoteaccount` are
advised not to use the Alembic migration and to instead follow the
upgrade guide.