feat: link-only remotes
* Added support for remote apps that can be set to "link only": i.e.
they can be connected to existing Invenio accounts but cannot be used to
create new accounts or log in to existing ones. As such, they are only
usable for third-party integrations rather than as identity providers.

* The `link_only` option is a key on the remote app config dictionary.
It does not affect remote accounts stored in the DB in any way, and can
be enabled or disabled at any time. By default it is false, so current
functionality is unaffected and these changes are fully
backward-compatible. When enabled, the remote is hidden on the login
page and signing in with it is prohibited on the backend with a
user-facing flash error message.

* If local accounts are disabled, users can disconnect all link-only
accounts freely but cannot disconnect the last non-link-only account.