fix(refresh): allow non-compliant access token response type
* GitHub and a small number of other implementations violate the RFC
6749 5.1 by returning a default response type of
`application/x-www-form-urlencoded` for the access token request,
despite `application/json` being mandatory. During the response parsing
in flask-oauthlib we therefore cannot identify the correct datatypes of
the response params, as they're all string.

* With the normal JSON type we can easily identify `expires_in` as an
int. To continue this behaviour, I added a simple check to convert a
string `expires_in` to an int. Without this, an error would be thrown
during authentication with these implementations.